requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi-bindings==25.1.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.7.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
realtime==2.22.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
